+++ cli/cli/Options/DotnetPathOption.cs
@@ -1,0 +1,10 @@
+using System.CommandLine;
+
+namespace cli;
+
+public class DotnetPathOption : Option<string>
+{
+	public DotnetPathOption() : base("--dotnet-path", "a custom location for dotnet")
+	{
+	}
+}